Quick fixes
tableList now looks like it should for when it's passed to innerJoin. Everything should be set up to have innerJoin take tableList/condition then return the joined table.

I also removed the line of code declaring any indexerror to be an invalid input line because that was... probably not a great idea in the first place.